backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
# cors is built into fastapi
